rename joingamecontainer to pregamecontainer
