+++ play/src/pages/pregame/HowToPlay.tsx
