+++ play/src/pages/pregame/SplashScreen.tsx
