+++ play/src/pages/pregame/howtoplayslides/HowToPlaySlide0Content.tsx
