+++ play/src/pages/pregame/howtoplayslides/HowToPlaySlide1Content.tsx
