+++ play/src/pages/pregame/howtoplayslides/HowToPlaySlide2Content.tsx
