+++ play/src/pages/pregame/howtoplayslides/HowToPlaySlide3Content.tsx
