+++ play/src/pages/pregame/howtoplayslides/HowToPlaySlide4Content.tsx
